% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{Infernape}
\alias{Infernape}
\title{Detect peaks from sorted bam files}
\usage{
Infernape(
  genome.ref,
  bam,
  batch.start = 1,
  batch.end = 10,
  ncores = 1,
  d = 31,
  h = 5,
  d.cut = 50,
  hr = 160,
  min.mode.prop = 0.05,
  min.mode.cutoff = 10,
  output.path,
  suffix,
  pas.reference.file,
  genome,
  pas.search.cut.1 = 0,
  pas.search.cut.2 = 300,
  polystretch_length = 13,
  max_mismatch = 1,
  motif.search.cut = 300,
  invert_strand = FALSE
)
}
\arguments{
\item{genome.ref}{Genome reference gtf file.}

\item{bam}{Sorted bam file.}

\item{batch.start}{ID of the first gene in this batch.}

\item{batch.end}{ID of the last gene in this batch.}

\item{ncores}{The number of cores.}

\item{d}{A parameter for smoothing.}

\item{h}{A parameter for smoothing.}

\item{d.cut}{A parameter for merging close peaks.}

\item{hr}{The estimated half range of peak mode - PAS interval. Used for annotating peak modes.}

\item{min.mode.prop}{Minimal relative threshold (wrt highest peak in 3'UTR region) to filter out small peaks.}

\item{min.mode.cutoff}{Minimal absolute coverage threshold to filter out small peaks.}

\item{output.path}{Output directory.}

\item{suffix}{Suffix as unique label for output table.}

\item{pas.reference.file}{Known PAS table}

\item{genome}{Full genome sequence. Required to be a S4 object.}

\item{pas.search.cut.1}{Distance (nt) from the peak mode to upstream boundary of the searching interval for PAS}

\item{pas.search.cut.2}{Distance (nt) from the peak mode to downstream boundary of the searching interval for PAS}

\item{polystretch_length}{Length of consecutive A sequence}

\item{max_mismatch}{Maximal tolerance of mismatch}

\item{motif.search.cut}{Window width for searching specified motifs}

\item{invert_strand}{default FALSE}
}
\value{
A peak table saved as [output.path]/peaks.[suffix].txt.
A peak annotation table saved as [output.path]/anno.[suffix].txt.
}
\description{
Detect peaks from sorted bam files
}
