% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main_apa_test.R
\name{Infernape_apa}
\alias{Infernape_apa}
\title{Differential APA testing}
\usage{
Infernape_apa(
  counts.dir,
  attr.file,
  anno.file,
  utr3.file,
  ctype.colname,
  base_grp,
  alt_grp = NULL,
  cut.low.pct = 0.05,
  cut.pval = 0.05,
  cut.MPRO = 0.2,
  test.type,
  out.dir
)
}
\arguments{
\item{counts.dir}{Directory for count matrix and corresponding peak site names and cell barcodes.}

\item{attr.file}{Directory for the cell attributes table.}

\item{anno.file}{Directory for the peak annotation table.}

\item{utr3.file}{Directory for the 3'UTR annotation table (mapping each 3'UTR region to a unique transcript ID).}

\item{ctype.colname}{Column name that defines cell groups in the cell attributes table.}

\item{base_grp}{Base cell group name.}

\item{alt_grp}{Alternative cell group name. If 'NULL', take all cell groups except base as alternative.}

\item{cut.low.pct}{Threshold for low expressing percentage. Default to be 0.05.}

\item{cut.pval}{Threshold for p-value. Default to be 0.05.}

\item{cut.MPRO}{Threshold for effect size (MPRO) filtering. Default to be 0.2.}

\item{test.type}{Type of testing. Value can be 'gene', 'utr', 'btw'.}

\item{out.dir}{Directory to save test results.}
}
\value{
Differential APA testing result table.
}
\description{
Differential APA testing
}
\examples{
ls()




}
