% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{peak_annotation}
\alias{peak_annotation}
\title{Annotate peaks' associated PAS and motifs}
\usage{
peak_annotation(
  peak.sites.file = peak.sites.file,
  pas.reference.file = pas.reference.file,
  output.file,
  genome = genome,
  pas.search.cut.1 = 0,
  pas.search.cut.2 = 300,
  polystretch_length = 13,
  max_mismatch = 1,
  motif.search.cut = 300,
  invert_strand = FALSE
)
}
\arguments{
\item{peak.sites.file}{Peak sites table generated by function peak_calling}

\item{pas.reference.file}{Known PAS table}

\item{output.file}{Output directory}

\item{genome}{Full genome sequence}

\item{pas.search.cut.1}{Distance (nt) from the peak mode to upstream boundary of the searching interval for PAS}

\item{pas.search.cut.2}{Distance (nt) from the peak mode to downstream boundary of the searching interval for PAS}

\item{polystretch_length}{Length of consecutive A sequence}

\item{max_mismatch}{Maximal tolerance of mismatch}

\item{motif.search.cut}{Window width for searching specified motifs}

\item{invert_strand}{default FALSE}
}
\value{
A peak annotation table.
}
\description{
Annotate peaks' associated PAS and motifs
}
\examples{
ls()


}
