% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_counting.R
\name{peak_counting}
\alias{peak_counting}
\title{Count reads for each peak site}
\usage{
peak_counting(
  bamfile = bamfile[i],
  whitelist.file = whitelist.file[i],
  peak.sites.file = peak.sites.file,
  ncores = ncores,
  CBtag = "CB",
  UMItag = "UB",
  start.cid = 1,
  end.cid = 26316
)
}
\arguments{
\item{bamfile}{BAM file.}

\item{whitelist.file}{Whitelist of cell barcodes.}

\item{peak.sites.file}{Peak site table after filtering.}

\item{ncores}{Number of cores.}

\item{CBtag}{Cell barcode tag in BAM file.}

\item{UMItag}{UMI barcode tag in BAM file.}

\item{start.cid}{The first peak cluster ID to analyze.}

\item{end.cid}{The last peak cluster ID to analyze.}
}
\value{
A peak site by cell UMI count matrix.
}
\description{
Count reads for each peak site
}
\examples{
ls()


}
