% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_filtering.R
\name{peak_filtering}
\alias{peak_filtering}
\title{Filter peaks}
\usage{
peak_filtering(anno.raw.file, q)
}
\arguments{
\item{anno.raw.file}{Peak annotation table.}

\item{q}{Vector of length 2 which defines the PAS searching window.}
}
\value{
A peak annotation table after filtering.
}
\description{
Filter peaks
}
\examples{
ls()

}
