% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_filtering.R
\name{rule_PAS_mode_dist}
\alias{rule_PAS_mode_dist}
\title{Adaptively determine filtering parameters using single PAS genes as negative controls.}
\usage{
rule_PAS_mode_dist(
  anno.raw.file,
  pas.reference.file,
  q.low,
  q.high,
  plot.output.dir
)
}
\arguments{
\item{anno.raw.file}{Peak annotation table.}

\item{pas.reference.file}{Known PAS table.}

\item{q.low}{Probability corresponding to quantile low bound}

\item{q.high}{Probability corresponding to quantile upper bound}

\item{plot.output.dir}{Directory to output plot.}
}
\value{
Quantile lower/upper bounds and a plot showing distribution of peak mode-PAS distance for negative control genes.
}
\description{
Adaptively determine filtering parameters using single PAS genes as negative controls.
}
\examples{
ls()


}
