% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_annotation.R
\name{peak_annotation}
\alias{peak_annotation}
\title{Annotate peaks' associated PAS and motifs}
\usage{
peak_annotation(
  peak.sites.file = peak.sites.file,
  pas.reference.file = pas.reference.file,
  output.file = anno.output.file,
  genome = genome,
  pas.search.cut.1 = 0,
  pas.search.cut.2 = 300,
  polystretch_length = 13,
  max_mismatch = 1,
  motif.search.cut = 300,
  invert_strand = FALSE
)
}
\arguments{
\item{peak.sites.file}{peak sites table generated by function peak_calling}

\item{invert_strand}{}
}
\description{
Annotate peaks' associated PAS and motifs
}
