% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_calling.R
\name{peak_calling}
\alias{peak_calling}
\title{Detect peaks from sorted bam files}
\usage{
peak_calling(genome.ref)
}
\arguments{
\item{genome.ref}{genome reference gtf file.}
}
\value{
message showing the total number of genes in genome reference.
}
\description{
Detect peaks from sorted bam files
}
