% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peak_counting.R
\name{peak_counting}
\alias{peak_counting}
\title{Count reads for each peak site}
\usage{
peak_counting(
  bamfile,
  whitelist.file,
  peak.sites.file,
  ncores = ncores,
  start.cid = NULL,
  end.cid = NULL
)
}
\arguments{
\item{bamfile}{BAM file.}

\item{whitelist.file}{Whitelist of cell barcodes.}

\item{peak.sites.file}{Peak site table after filtering.}

\item{ncores}{Number of cores.}

\item{start.cid}{The first peak cluster ID to analyze.}

\item{end.cid}{The last peak cluster ID to analyze.}
}
\value{
A peak site by cell UMI count matrix (sparse).
}
\description{
Count reads for each peak site
}
\examples{
ls()


}
